Deskbar: new single line mini-mode
Mini mode is when Deskbar is in the 4 corners with the application
menu tucked away into the leaf menu bar. This commit switches mini
mode from putting the clock and replicants below the leaf menu to
attaching them to the side of leaf menu forming a single line.
The main advantage of this is that if Deskbar is in the top right
corner then you can maximize a window without losing any access to
Deskbar or any screen real-estate with any other corner losing
screen real-estate similar to horizontal mode.

Replace team icon with HVIF version kindly donated by meanwhile.
Scale icon size according to font size.

Switch order of leaf and app list menus on the right side so that
the leaf menu is the one on the screen edge.

Update SetSizeLimits for mini mode - same as horizontal mode but
doesn't need to scale vertically (except with font size.)

Add BarView pointer to TimeView. Ditch fVertical and SetOrientation
in TTimeView and get the information from BarView instead. This way
we don't have to keep the state in sync.

Truncate time in vertical expando mode, no longer truncate time in
vertical mini mode because now the time can expand itself
horizontally without limit (within reason.)

Pass BarView pointer to TTeamMenu constructor and use it instead of
getting the pointer in a roundabout way from be_app.

Adjust the width of the team menu for larger icon sizes. If app
names are hidden limit to half of min width so the menu doesn't
appear too short. SetMaxContentSize to width, this fixes a bug where
the menu item wasn't quite the right size for the menu.

Switch to mini-mode using 1/3 screen left and right of screen instead
of using 1/6. Mini mode is now wider and this change makes switching
between mini mode and horizontal mode more reliable.
 * Rename frame to screenFrame here for clarity.

Eliminate kMiniHeight constant, the mini-mode height is now based
on the height of a window tab which changes with be_bold_font size.
Limit maximum height of replicants by be_bold_font size as well.
Also limit team menu and leaf menu icon heights by be_bold_font size.

Set Menu Bar height to window tab height in vertical mode. This is
so that when you maximize a window the tab size will match Deskbar.
the menu bar height is also a few pixels taller (25px vs 21px) than
it was in vertical expando mode at the default 12px font size. The
change was made to match the tab height, 21px was the height
of a window tab on BeOS R5 on the default decorator.

Adjust leaf menu position so that it continues to be cut-off at the
bottom in vertical mode.

Constrain leaf bitmap to its menu item frame as it can bleed out
into the adjacent team menu item at larger font sizes.

Implements bug #5876 step 1.

Change-Id: Id9ddd60c997a785184208ba02938bee1416aeae9