BDeskbar API: Add recent count getter and setter methods
GetRecentCounts and SetRecentCounts use varargs for future expansion.
First param is the number of counts to get or set. The remaining params
get or set recent applications, folders, and documents respectively.
Set a value to NULL for getter or -1 for setter to ignore the value.
This is useful if you only want to get or set the 2nd or 3rd count.
Also added convinience methods which get or set a single recent count.

Grab recent folder count from Background prefs instead of hardcoding 10.

In hrev52033 David Murphy added a recent folders feature to Background
preferences. This updates it to use the newly accessible setting from
the Deskbar API instead of hardcoding a value of 10.