Update patch set 2
Patch Set 2:

Responding here since I'm not aware of a ticket for this. When I worked on Backgrounds previously, the feedback I received from others on the Haiku team was that applications should maintain their own settings with regards to "recent folder counts" or other settings. I actually had code that retrieved the setting (much less elegantly than what you have here) from Deskbar and used it to drive Backgrounds recent folders but I removed it since there was a general sentiment that this was a "bad idea". (see https://dev.haiku-os.org/ticket/5026, also https://dev.haiku-os.org/ticket/14181)

I guess the fundamental question here remains - is it intuitive to a user that a "recent folder" count setting in the Deskbar is going to translate to other applications? If so, how should this relationship be established graphically?  And is the existing Deskbar setting the right one to use to drive other applications?

Patch-set: 2