Update patch set 2
Patch Set 2:

I may have inadvertantly stepped on a bee's nest here. Consider this my vote for using the Deskbar recent folder setting in Backgrounds. Your commit in Backgrounds was very good just missing this piece and that was not your fault, BDeskbar API is incomplete and needs to expose more of Deskbar's settings to apps.

Patch-set: 2