Update patch set 3
Patch Set 3:

waddlesplash, I forgot to address your question about the "Fixme", so here's my response to that:

When reading that fixme, I thought it was referring to the line below where the 'fState' instance variable was being set. The comment was referring to problems with keyboard navigation during the attach - which is the very issue I solved in AttachedToWindow(), so from my standpoint I did resolve the fixme. If you disagree and think it's referring to the above line with the _setStickMyMode, I can certainly restore the comment, but that's not how I understood it. In my experience comments refer to things below not things above

Patch-set: 3
Label: Code-Review=0