Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Is it intuitive to a user that a "recent folder" count setting in the Deskbar is going to translate to other applications?
> 
> Sharing the Deskbar recent folders setting with Backgrounds falls into the category of "Sane defaults, not maximal configurability".

My opinion, such as it is, (being new to the project) is that having a centralized setting for "recent items" makes a lot of sense. However, I think the best solution would be for it to be a separate setting, not a part of deskbar. Rather it would be a centralized setting used by Deskbar, tracker, and all other Haiku built-in apps and preferences. Right now I believe Deskbar and Tracker have separate versions of these recent items settings

Patch-set: 2
Reviewer: Gerrit User 1000059 <1000059@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0