Update patch set 3
Patch Set 3:

(1 comment)

Yes, the FIXME definitely refers to the line above it. My guess is when keyboard-navigating, we force the menu to be "sticky" (not closed on mouse button release).
We don't use much non-sticky menus, but we should test keyboard navigation on them as well, I guess.

Patch-set: 3
Label: Code-Review=0