Update patch set 2
Patch Set 2:

You can check src/servers/app/drawing/AlphaMask.cpp, which uses an UtilityBitmap (a ServerBitmap subclass intended for such uses - not linked to a client BBitmap or so) and then draws on it (see _RenderSource function).

Eventually it gets to creating a DrawState using the bitmap, which can be used to draw things, with an API similar to BView.

Patch-set: 2
Label: Code-Review=0