Visually align shortcuts in menus
In a menu, we use the right side both for submenu arrows and shortcuts.
As a result, when an entry has both a shortcut and a submenu, its
shortcut is not aligned with others, and this does not look so nice.

TODO (axeld in #8900):
"I would say the spacing should only appear if there is any sub menu,
otherwise, I could imagine this won't look as good. I like the changed
alignment better in any case, though."