Update patch set 4
Patch Set 4: Code-Review+1

I thought maybe DrawBitmap was doing some sort of anti-aliasing... but no, it turns out that all the bitmaps I was drawing into were 1px too big (e.g 8x8 instead of 7x7). Patch 4 functions identically to patch 3 but uses ServerBitmaps and DrawBitmap instead of defining a bunch of BRects and calling DrawRegion. It is much easier to visualize what the bitmaps are doing than it was to contemplate BRegions.

Patch-set: 4