Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This is not quite there yet but I may have to respond with a patch. The TeamMenuItem.cpp changes don't belong as explained inline. You are declaring but not using kHPad in MenuItem.cpp, and that constant only makes sense in Deskbar anyway. The tint color needs to be at least B_DARKEN_4_TINT. Actually should be between 4 and MAX maybe around 1.75. B_DARKEN_3_TINT is right out. Using text color might be a good idea too so that when we apply blue on white patch you hate the arrow will be drawn white instead of black. Also I'd like the left and right menu margins to match so that we can apply Ryan Leavengood's patch and the menu items will look nice and balanced.

Patch-set: 1
Label: Code-Review=-1