Update patch set 1
Patch Set 1:

The spacing is ok, the left side is used in the same way for the checkmark (which we could shift one pixel to the right)

I only tested this with my small font size, maybe it does not look as good at other sizes. We probably need to review the whole layouting code?

Patch-set: 1
Label: Code-Review=0