Update patch set 2
Patch Set 2:

We should make a decision on the padding (earlier version of the patch had the arrow slightly more to the left). Right now it is the same as the checkmark on the left side (very close to the edge as well). Spacing between shortcuts and the arrow may be slightly reduced as well.

And no, this probably won't work with a dark background, BControlLook does not allow specifying a color here, only a tint to derive from the base. DeskBar has the same problem, then, and probably scrollbar arrows and notifications fold/unfold as well (they all use the same code).

Patch-set: 2
Label: Code-Review=0