Draw ScrollBar arrows in BScrollBar using HaikuControlLook
This commit seeks to restore ScrollBar functionality by bring the
arrow position code back into BScrollBar and FakeScrollBar.
This complicates the BScrollBar::Draw() code but has the benefit
of working more in line with how WebPositive expects.

Kept virtual method count the same in BControlLook by taking out
the DrawScrollBar method and replacing it with DrawScrollBarBackground
which is the method that WebPositive is expecting. Keep
DrawScrollBarThumb for drawing the thumb. It now draws just the thumb
and the non-thumb area is drawn using DrawScrollBarBackground.
The arrows are drawn in BScrollBar by calling a combination of
DrawButtonBackground and DrawArrowShape.