BWindow: Zoom maximize with shift held down
In hrev51623 the default zoom behavior was changed to respect Deskbar,
that is to say that the Window will zoom to screen area minus the area
taken up by Deskbar.

Some have complained that they liked the old maximize behavior. This
commit updates the default zoom behavior so that if you shift+Zoom it
will maximize the window ignorning Deskbar. That is to say that the
window will resize itself to take up the entire screen frame.

Note that if you override Zoom in your app then shift+Zoom will not
work. This only works for apps that use the default BWindow::Zoom.