Update patch set 1
Patch Set 1:

My display is quite low-res (1360x768) and after some time experimenting with the "avoid deskbar" mode, I would rather have this by default, not hidden under a key shortcut.

"Avoid the deskbar" should be in force only when the deskbar is in horizontal mode (at top or bottom of screen), and never when it is in vertical mode (the proper thing to do there is use deskbar autoraise, I think).

Patch-set: 1
Label: Code-Review=0