Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

Andrew, it's a very valid concern. I'm not sure this more complex version of the patch provides enough benefit to justify both the extra code as well as the fact that we will be using more memory to load all the bitmaps. After all, a HVIF scaled down to 16x16 looks basically as good as a native 16x16 icon.  That is why I created 2 patches - this more "comprehensive" one, and the simpler one that just makes HaikuDepot icon sizes fixed in the 2 list views, and uses BView scaling to draw it at the requested size.  I'm OK with either one going through. 

The only point I think in favor of this version of this patch is that it does expose all the package data that's available in our Haiku depot. If we don't care about any icon except the best resolution, then why store them at all in the package? Just have developers upload their best resolution and we take care of the rest. Also I think HaikuDepot is already keeping a lot of other stuff in memory, and these bitmaps aren't really all that big. By the time we need to refactor the bitmaps we may also need to refactor other aspects of what's being loaded into memory.

Ultimately it's up to you and the rest of Haiku team. I'm just glad I can get rid of a bug that makes HaikuDepot look quite strange!

Patch-set: 3