Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

The (x,y) in this case are the origin coordinates where the view will be drawn. I was thinking that the view will be explicitly 16x16 so I added 16.  The BRect above of course is starting at 0 so we add 15 there to account for the zero indexing... You may be right here, I'm still getting used to the drawing code. I'll do some testing tonight

Patch-set: 2