Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > The only point I think in favor of this version of this patch is that it does expose all the package data that's available in our Haiku depot. If we don't care about any icon except the best resolution, then why store them at all in the package?
> 
> I think both forms of the patch are worthwhile; we can use the scaling if there is only one size / HVIF available, and if there is only a bitmap icon. Is that a good option? Or should we really only merge one?

The first patch (347) is required to fix the bug, that's why that portion is included in both patches.  The extra stuff in this patch (353) is really around adding access to all available icons of a package, and choosing the correct to display in each circumstances instead of using scaling.  So 353 includes 347, but we could instead opt to just merge 347 and abandon 353. It just depends on if we think making all the icons available in HaikuDepot is useful or not. Before the code was simply picking the best resolution icon and using that in all circumstances, but obviously this requires scaling the icon to the size of the BView (which is what patch 347 does).

Patch-set: 3