Update patch set 2
Patch Set 2:

waddlesplash, Andrew,
I can confirm that the original value of 16 is correct. If you use a value of 15, a portion of the bitmap is cutoff. I was able to visibly confirm this looking at the icons as drawn in the package list view.  So this code is good to go.

Patch-set: 2
Reviewer: Gerrit User 1000059 <1000059@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0