use-clang: now 75% more gcc-less
These changes allow you to choose the local clang compiler to
cross-compile Haiku without the need for build-cross-tools.

The prefix works differently on clang vs gcc due to the nature
of the clang tool namings. Technically we could adapt the same
prefix for the gcc tools, but (triplet)-gcc is likely prefered
over (triplet)/gcc for clairiy.

Change-Id: I8715b4ed680ef19677f94713fbb511d46566b77a