Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Besides the issues mentionned in comments, you should also take care of not saving the setting when it is not supported by the current driver or display. This is to make sure booting in VESA does not accidentally reset the setting to an unexpected value, for example.

Also something I did not check: when starting the screen preferences, it should always get the current brightness value from the video card, not from the file. What if there is a mismatch? Typically, when we get to implementing keyboard shortcuts to change the brightness, how will these be handled?

Patch-set: 1
Label: Code-Review=-1