Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Still two coding style violations.

And I'm now not completely happy with the design and the way the settings are stored (sorry for not identifying the problem earlier).

The "Screen_data" settings file is storing only the window position so far. It does not seem to be the right place to store this information. Why is it not stored with the display modes?

The reason for this is later on there will be other ways to change the brightness (for example, using the screenmode command - which can then be attached to a keyboard shortcut or so), and these alternate ways would then need to mess with the Screen_data settings file as well.

Moreover, the API for setting the brightness was put in BScreen, because when we start handling multiple displays, we want a different brightness level for each display. So it makes little sense to store only a single brightness setting, and the brightness should be attached to each display (just like the video mode).

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1