Update patch set 11
Patch Set 11: Code-Review-1

(4 comments)

Some coding style violations again, but not a lot.

Once these are fixed, it would be nice if someone else could review the code as well, because I discussed the design a lot and it is nice to have an external eye on it. So when you have fixed the remaining issues, ask for review on IRC or mailing lists.

Patch-set: 11
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1