Update patch set 1
Patch Set 1:

not 100% on this one.

FDT code has been kind of spread out and hacked into everything from a u-boot perspective. We end up with openfirmware using u-boot things which is weird.
We might need FDT on EFI for arm64 as well (unsure)

Waddlesplash had the idea of putting it into a bus_manager like packagefs.  The oddity is we need FDT in the following places:

  * static haiku_loader -- mmu mapping, serial, framebuffer, and other misc things
  * static kernel -- serial uart and early framebuffer stuff
  * add-ons fdt bus -- everything else like usb, etc

Patch-set: 1
Label: Code-Review=0