Update patch set 2
Patch Set 2:

technically there is a fdt bus_manager already started (which is why it didn't show up in the commits)

https://git.haiku-os.org/haiku/tree/src/add-ons/kernel/bus_managers/fdt/fdt.cpp#n160

In these commits, i'm moving fdt support code used in the bootloader and kernel code out of the various kernel/bootloader directories and putting it into a common spot under the existing fdt bus_manager. The fdt bus_manager now generates a bootloader and kernel "helper library" which give each access to fdt libraries and helpers.

Historically the FDT stuff was implemented with a focus on u-boot (and 
 then openfirmware was hacked in).  This makes everything more "generic" (without using the "generic" platform since this code really isn't "generic" to all platforms... just common.)  The goal is to try and consolidate this stuff down and encourage the use of an FDT "bus" in the future for less-essential kernel fdt access needs.

We're always going to need "early" access to fdt however in the bootloader and  kernel (esp for serial debugging)

Patch-set: 2