Update patch set 2
Patch Set 2:

I think you missed my comment on the ticket.

For serial access, the FDT is not of much use because you can't just grab the first serial port and hope for the best. So we need to use the kernel settings to find where the serial port is.

The kernel setting could be either an hardcoded address, or an FDT path (likewise on x86 it is either an hardcoded IO port, or a BIOS port number). If using an hardcoded address, no FDT lookup is needed at all.

We definitely need the bus manager, which allows each device driver to map the memory it needs (just like the PCI bus allows device drivers to map the memory or IO space they need - the PCI bus itself does not take care of it).

Eventually we need to integrate this FDT bus with the device manager, and have device drivers handle either FDT or PCI (it would help if they both provide a common API).

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0