Update patch set 2
Patch Set 2:

Well, the goal here is to run the tests automatically on buildbots, which we can't if one of the test ends up crashing the bot.
So, the fist step is getting this merged and the buildbot set up. The second step will of course be fixing the problems found and re-enabling the tests.

Patch-set: 2
Label: Code-Review=0