Update patch set 1
Patch Set 1:

I’m pretty sure this will still work with the build of libpackage_build.so. During it’s build, FetchFileJob.cpp will simply not be built, and all of it’s use cases are surrounded with a check that the library is being built for the Haiku platform, but at the same time, not the Haiku host platform (Defined when building any *_build.so libraries).

Patch-set: 1
Label: Code-Review=0