Update patch set 1
Patch Set 1:

I had a look at the code involved for drawing a string in a BPicture.
While a comment in ServerPicture.cpp draw_string() says that the draw_string command is prepended with a set_pen_location command (in the BPicture data), we don't do that in any place.
I think that the correct fix for this problem has to be done in ServerWindow.cpp: 
line 3562, case AS_DRAW_STRING:

I think that a "WriteSetPenLocation()" with the current view location should fix this problem correctly.

Patch-set: 1