Update patch set 1
Patch Set 1:

I don't think it would be possible to update the pen location properly. BPicture does not know the font metrics, it should just store commands to be replayed at a later time.

So, when recording we should not rely on anything from the current view state (pen location, font size, ...) and end up storing that data in the BPicture.

This probably means changing the app_server interface so the pen location is fully managed on app_server side, by introducing a new "draw at current pen location, whatever it is" call, and not reading the position from app_server and pushing it again.

Patch-set: 1