Update patch set 2
Patch Set 2:

> > > Thanks for the quick review!
> > > One question: will someone bac port this into r1beta branch, or what's the preferred style in Haiku?
> > 
> > Someone may cherrypick this commit onto the beta-branch. However, since it is relevant only when building Haiku yourself, I guess there is no need to apply it to the beta branch. The only exception would be if someone has the beta-branch checked out for some reason instead of master. But perhaps I am missing something?
> 
> I see. I was just wondering whether there were formal rules, which patches to include in a release candidate branch and which not. Fully fine as it is now. Thanks for your help.

In the past, the policy was that there is a "release master", a person who is in charge of deciding which patches to merge from the master onto the release branch. I guess the idea is to do that only for patches that appear to have no risk of introducing new problems. So mostly only bug-fixes will be cherry-picked. But sometimes even a bug-fix patch may be complex enough to hold off on it.

Patch-set: 2