Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> Why do we have to resort to a settings file here? Why do these settings not show up in media preferences?

I suppose, because nobody submitted code doing that. There's been a discussion at the linked ticket if these settings should be at all exposed in the Media prefs, or if it could be calculated to use the right values for the right job, or if the "Realtime checkbox" should be introduced...

In any case, the patch has been there for 6 years, last discussion was 3 years ago. It doesn't look like the Media prefs will be changed any time soon.
Meanwhile, with this patch, users have at least a work-around to use Haiku for live music. Inconvenient, true, but better than not being able to use Haiku at all for this task. And the people that don't care about it aren't impacted at all, just like with the already existing optional settings for auich.

Patch-set: 4