Update patch set 5
Patch Set 5:

And is there a problem with just making the default buffer size smaller? Having to tweak settings files for this is very not Haiku-ish. Either we go the full way and make the settings visible in media prefs (the driver can expose whatever it wants there), or we just settle on sane defaults and don't add a setting at all.

The half-solution of adding a settings file, but then not showing it to users is not convincing.

Patch-set: 5
Label: Code-Review=0