Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Possibly, but it's an involved process to change what branch Pootle pulls strings from. That is the concern.

It's not really that involved. It would be enough to do a `git checkout` the other branch on the pootle repository and have it create the templates from there and push the strings to there.

In any case, the master branch does not have to be frozen anymore, since the strings are not updated automatically at this time. Furthermore, the deadline for string translations for beta 1 has passed. This means that the most recent commits are the last ones for beta 1.

Patch-set: 3
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0