Update patch set 1
Patch Set 1:

This works, but the point of adding DrawStringWithOffset was to speed up the operation by sending the whole string to app_server in one go.

The speedup will be lost here if I'm not mistaken?

Well, at least we get the thing working, then we can see about making it faster.

Patch-set: 1