Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I'm not so much in favor of implementing it like this. Solutions like these tend to stick. Why not do it right? What is the difficulty?

It's not really a matter of how difficult it is, it's more a matter of if it's really needed. The bpicture protocol is somewhat simpler than the app server one, and while we are definitely in control of it, I think that we should avoid changing it if it's not necessary. For example, there is no "fill/stroke triangle", no "constrain clipping region", and no "invert rect": they are implemented using other simpler ops.
Implementing this draw string version this way, existing users (there are probably none, or very few, though) don't even have to implement a new op, and we keep the bpicture format simpler.

Patch-set: 1