Update patch set 1
Patch Set 1:

As I mentioned in the ticket, this route usually has incorrect font kerning. It may look alright with our default settings, but if you enable LCD subpixel hinting, doesn't it look worse? Already there is something else in appserver that does this per character for something, so a lot of strings already look terrible, but ones drawn with this API do not presently IIRC.

So the correct solution would be better, I think.

Patch-set: 1