Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I would prefer to make the ioctl() call in devfs_sync() and devfs_fsync(), and then simply call that.
> Well, this ioctl is valid for block drives only, right? So calling it unconditionally in devfs would not be correct. Or does devfs know what devices are block, etc.?

It doesn't have to; if you understand the B_FLUSH_DRIVE_CACHE, you perform it, otherwise, you don't. It's the same as before, just that the knowledge is moved closer to where it belongs.

And you'll actually get a working fsync() functionality for 'dd' on the way.

> Duplicate syncs will happen under that model too.

Yes, that's a completely independent problem.

Patch-set: 2