Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> https://www.haiku-os.org/legacy-docs/bebook/BRect_Overview.html
> 
> "You would expect a BRect defined thus…:
> 
> BRect rect(0, 0, 3, 3);
> 
> …to have a width of 3.0 and a height of 3.0. These, indeed, are the values returned by the Width() and Height() functions."
> 
> So according to the docs if I want to make the rectangle wider by 16 pixels I need to add 15 to it? What the hell?

No, if you want to make it /wider/ by 16 pixels, you need to add 16 to the width. But that is not the question here. The culprit is what these values mean. Since they are floating point values, you would expect they mean "pixel coordinates". You would think a left side of "0" means the left side of the pixel at index 0. And a value of "16" would mean the right edge of the pixel at index 15. So an "enclosing" rectangle. However, what they mean instead is "pixel indices". Yes! A BRect(0, 0, 15, 15) covers the pixels with the indeces 0 through 15. Which ammount to 16 pixels. This is very confusing and not at all how other APIs work, and you are not the first now will you be the last to be confused by this, but it is how the Be API works. Please read the chapter on BView drawing, which covers this.

You can also read through the BRect code. A lot of it will make more sense. For example, to make a BRect that covers one pixel, you give it the same value for left and right edge. So BRect(0,0,0,0) covers the pixel at index 0. It is not an empty rectangle, like one would expect! That's why IsValid() checks whether right is /smaller/ than left and so on.

If you still don't believe me, you can also use Magnify and just count the pixels in HaikuDepot that are covered by an icon now. It should now be 15.

Best regards,
-Stephan

Patch-set: 2