Update patch set 2
Patch Set 2:

OK, I get it now. Indeed the (x, y, x + 16, y + 16) is the problem here. It does not make the rect wider, it initializes it so it should have 15 there.

But is there any good reason to do it this way? All this does is create confusion and weird corner cases like this or two BRects (0, 0, 2, 2) and (3, 3, 5, 5) apparently being 2 pixels wide and 1 pixel apart when in reality they're 3 pixels wide and there is no gap between them. What is the reasoning for this insanity?

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0