Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> OK, I get it now. Indeed the (x, y, x + 16, y + 16) is the problem here. It does not make the rect wider, it initializes it so it should have 15 there.
> 
> But is there any good reason to do it this way? All this does is create confusion and weird corner cases like this or two BRects (0, 0, 2, 2) and (3, 3, 5, 5) apparently being 2 pixels wide and 1 pixel apart when in reality they're 3 pixels wide and there is no gap between them. What is the reasoning for this insanity?

You would need to ask the Be engineers. We inherited this insanity. You can see their line of thinking somewhat in the BRect documentation which you have pointed to. As I said, you need to think of the coordinates as pixel indeces. However, that does not help with the insanity of BRect.Width() simply returning a wrong value. Or at least a completely unexpected value. It would be somewhat more OK, if it was implemented as "right - left + 1". But it isn't, which makes the API inconsistent in itself. But we cannot change it without breaking existing apps.

Patch-set: 2