Update patch set 2
Patch Set 2:

OK, maybe some more elaboration in credit of the Be engineers... as is documented in the BeBook, a coordinate of (0, 0) means the *center* of the pixel at index [0, 0] in a buffer. Not the top-left corner of that pixel as in other APIs. The top left corner would be "-0.5, -0.5" and the bottom right corner of the pixel would be "0.5, 0.5". The reason why that may be helpful is when you consider a transition from an integer based coordinate system to a floating point one. In 1993, you have all this code which draws using integer pixel coordinates. But the Be drawing API introduces floating point coordinates. Now, if you want to draw a line from one pixel to another, in a traditional integer coordinate system, you fully hit a pixel at (0, 0). In a floating point coordinate system, you need to consider the line width of 1 and imagine a stroke with round or square caps. In order not to change the existing drawing code, it helps when (0, 0) means the center of the lixel, not the left-top, since then no matter where the line is directed at with a (0, 0) start point centered on the pixel, the starting pixel is always covered like it was with the integer based drawing API.

That's all fine and good, but BRect::Width() is still returning an unexpected value.

Patch-set: 2