Update patch set 2
Patch Set 2:

It would be nice if you actually left these changes here for a while so other people can actually comment.

I indeed did that change locally, but while reviewing I realized that it would possibly make more sense to let the overall preparation not fail in such cases at all. The wording of the message at least suggests that it was meant to be ignored. It is a tradeoff that could also be determined by numbers, i.e. how many interrupts are routed vs. how many were assigned by the firmware. If there are more routed interrupts with the IO-APIC than without, it should probably be used.

We must also ensure that the 0xff interrupt line is reset to 0 in such cases, as otherwise the exact problem the Linux commit tries to avoid may come up (i.e. a driver trying to use 255 as the interrupt). I have not checked whether this already happens implicitly. The reason why the Linux commit only happened in 2016 is probably that before the widespread use of MSIs, it was extremely unlikely that you would collide on such a high numbered interrupt.

Please also adjust the debug output to print the actual value, so that we can tell the cases apart without having to look up the real value in the PCI dump.

I've seen that the system in #14504 is also impacted by this (there it is an unrouted SMBus device). So that one should be retested as well.

Patch-set: 2
Reviewer: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0